define the trovefiles.path column as BYTEA for PostgreSQL. This is required since declaring it as a UTF8 VARCHAR makes PostgreSQL check the validity of the UTF8 strings being inserted into the path column, which is not something we want.
For compatibility with 1.0 line, we leave the mysql and sqlite schema
definitions the same, since we can not go through a schema migration.
This works okay on mysql for now because we use latin1 with a
latin1_bin collation.

For conary 1.1 it is better to go through a schema migration to take
the TroveFiles.path column out of the constraints of matching a
particular character set; paths should be just a string of bytes with
no particular locale meaning attached to them.