Added downloadUpdate method. (CNY-1300)
There is now an optional step that you can run, downloadUpdate, that will
download the changesets and save them on the disk. A reference to them is kept
in the update job object, and it is properly preserverd over freezing/thawing.

If applyUpdate is called on an update object for which the changesets were
already downloaded, the download step is skipped.